feat: 메타데이터 추가
